Improve motion sensor calibration and quaternion handling
Refactored motion data processing to use correct buffer offsets and improved calibration factor calculations for accelerometer and gyroscope. Changed accelerometer bias types to float and updated Madgwick filter usage for more accurate quaternion orientation. Adjusted device polling interval for higher update rate and removed unnecessary HID queue flush. Cleaned up code and comments for clarity.

+++ Source/WindowsDualsense_ds5w/Public/Core/Types/Structs/Config/GamepadCalibration.h
@@ -29,9 +29,9 @@ struct FGamepadCalibration
 	int16 GyroBiasY = 0;
 	int16 GyroBiasZ = 0;
 
-	int16 AccelBiasX = 0;
-	int16 AccelBiasY = 0;
-	int16 AccelBiasZ = 0;
+	float AccelBiasX = 0;
+	float AccelBiasY = 0;
+	float AccelBiasZ = 0;
 
 	float GyroFactorX = 1.0f;
 	float GyroFactorY = 1.0f;
